Some buffs to the Altar upgrades (#15)
* Some buffs to the Altar upgrades

These are relatively small to mild, but should suffice for some nice buffs. Starting off relatively small for now to test the waters.

* Readjust some values

Reconfirmation:  `this.consumptionMultiplier = (float) (0.25 * upgrades.getSpeedUpgrades());` seems to be for Orb input? And `this.efficiencyMultiplier = (float) Math.pow(0.80, upgrades.getSpeedUpgrades());` seems to be for everything else like normal recipes.

* Update dependencies.gradle

* Update dependencies.gradle

Co-authored-by: Martin Robertz <dream-master@gmx.net>